# docs/regression-baseline/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Base64 Convertor API (Baseline)",
    "version": "0.0.1-SNAPSHOT",
    "description": "Hand-authored OpenAPI baseline. No OpenAPI/Swagger generator (e.g. springdoc-openapi) is present in pom.xml and no /v3/api-docs endpoint exists in this application, so this document was manually authored from the controller source and live baseline captures in docs/regression-baseline/, rather than exported from a running generator. Treat this as a snapshot for regression comparison, not a live contract."
  },
  "servers": [{ "url": "http://localhost:8080" }],
  "paths": {
...
    "/api/files/convert": {
      "post": {
        "summary": "Batch URL to Base64 conversion",
        "requestBody": { "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/FileConvertRequest" } } } } },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/FileConvertResponse" } } } } }, "400": { "description": "Bad Request" } }
      }
    },
...
    "/api/files/pdf/send": {
      "post": {
        "summary": "Fetch file and forward to target chat/message API",
        "requestBody": { "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/PdfRequest" } } } } },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/PdfResponse" } } } } } }
      }
    },
    "/api/files/pdf/convert/base64": {
      "post": {
        "summary": "Fetch file, return Base64",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PdfBase64Request" } } } },
        "responses": { "200": { "description": "OK" }, "502": { "description": "Bad Gateway" }, "500": { "description": "Internal Server Error" } }
      }
    },
    "/api/files/pdf/convert/base64dynamic": {
      "post": {
        "summary": "Flexible fetch (any method/headers) with validation",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PdfBase64RequestDynamic" } } } },
        "responses": { "200": { "description": "OK (status field in body may be SUCCESS/FAILED/FILE_TOO_LARGE)" } }
      }
    },
    "/api/files/pdf/single": {
      "post": {
        "summary": "Fetch + optional forward, single item",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PdfRequest" } } } },
        "responses": { "200": { "description": "OK" }, "502": { "description": "Bad Gateway" }, "500": { "description": "Internal Server Error" } }
      }
    },
...
  },
  "components": {
    "schemas": {
      "FileConvertRequest": { "type": "object", "properties": { "url": {"type":"string"}, "fileName": {"type":"string"}, "caption": {"type":"string"}, "mimeType": {"type":"string"}, "type": {"type":"string"} }, "required": ["url","mimeType","type"] },
      "FileConvertResponse": { "type": "object", "properties": { "processingId": {"type":"string","nullable":true}, "fileName": {"type":"string"}, "mimeType": {"type":"string"}, "type": {"type":"string"}, "base64Data": {"type":"string","nullable":true}, "success": {"type":"boolean"}, "message": {"type":"string"} } },
      "PdfRequest": { "type": "object", "properties": { "url": {"type":"string"}, "cookie": {"type":"string"}, "payload": {"type":"string"}, "target_url": {"type":"string"}, "target_auth": {"type":"string"}, "message": {"type":"object"}, "metaData": {"type":"object"} } },
      "PdfResponse": { "type": "object", "properties": { "fileName": {"type":"string"}, "status": {"type":"string"}, "base64": {"type":"string","nullable":true}, "mimeType": {"type":"string","nullable":true}, "size": {"type":"integer"} } },
      "PdfBase64Request": { "type": "object", "properties": { "url": {"type":"string"}, "cookie": {"type":"string"}, "payload": {"type":"string"} }, "required": ["url"] },
      "PdfBase64RequestDynamic": { "type": "object", "properties": { "url": {"type":"string"}, "cookie": {"type":"string"}, "httpMethod": {"type":"string"}, "payload": {"type":"object"}, "authorization": {"type":"string"} }, "required": ["url"] },
...
    }
  }
}